Move GraphQueryDescription to a separate query package
